+++ CHANGELOG.md
@@ -5,11 +5,37 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.43.2] - 2022-11-23
+## [3.44.4] - 2022-11-25
 
 ### Fixed
 
 - voice control is properly generated for selected language.
+
+## [3.44.3] - 2022-11-25
+
+### Fixed
+
+- added missing pageId at link translator
+
+## [3.44.2] - 2022-11-25
+
+### Fixed
+
+- added missing solution for tracking events `BotApp#trackEvent()`
+
+## [3.43.1] - 2022-11-24
+
+### Fixed
+
+- problem with `onInteractionHandler` interface
+
+## [3.44.0] - 2022-11-24
+
+### Added
+
+- `onInteractionHandler` function as a Processor's "interaction" event handler
+- `GA4()` class as a new GA4 adapter
+- `BotApp#registerAnalyticsStorage()` - method simplifies attaching analytics storage
 
 ## [3.43.1] - 2022-11-17